+++ packages/gestalt/src/Switch.tsx
@@ -1,8 +1,9 @@
-import { useState } from 'react';
 import classnames from 'classnames';
 import focusStyles from './Focus.css';
 import styles from './Switch.css';
 import useFocusVisible from './useFocusVisible';
+import useInExperiment from './useInExperiment';
+import useInteractiveStates from './utils/useInteractiveStates';
 
 type Props = {
   /**
@@ -48,7 +49,10 @@ export default function Switch({
   switched = false,
   dataTestId,
 }: Props) {
-  const [focused, setFocused] = useState(false);
+  const isInVRExperiment = useInExperiment({
+    webExperimentName: 'web_gestalt_visualRefresh',
+    mwebExperimentName: 'web_gestalt_visualRefresh',
+  });
 
   const handleChange: (event: React.ChangeEvent<HTMLInputElement>) => void = (
     event: React.ChangeEvent<HTMLInputElement>,
@@ -57,20 +61,33 @@ export default function Switch({
     onChange({ event, value: checked });
   };
 
+  const { handleOnFocus, handleOnBlur, isFocused } = useInteractiveStates();
   const { isFocusVisible } = useFocusVisible();
 
   const switchStyles = classnames(styles.switch, {
-    [focusStyles.accessibilityOutlineFocus]: focused && isFocusVisible,
+    [focusStyles.accessibilityOutlineFocus]: isFocused && isFocusVisible && !isInVRExperiment,
+    [styles.focus]: isFocused && isFocusVisible && isInVRExperiment,
     [styles.disabledSelected]: disabled && switched,
-    [styles.disabled]: disabled && !switched,
+    [styles.disabledUnselected]: disabled && !switched,
     [styles.enabledSelected]: !disabled && switched,
-    [styles.enabled]: !disabled && !switched,
+    [styles.enabledUnselected]:
+      !disabled && !switched && !(isFocused && isFocusVisible && isInVRExperiment),
+    [styles.borderColorTransition]: !isInVRExperiment,
   });
 
   const sliderStyles = classnames(
     styles.slider,
     switched ? styles.sliderRight : styles.sliderLeft,
     !disabled ? styles.sliderDark : styles.sliderLight,
+  );
+
+  const sliderVrStyles = classnames(
+    {
+      [styles.sliderVrRight]: switched,
+      [styles.sliderVrLeft]: !switched,
+      [styles.disabledSlider]: disabled && !switched,
+    },
+    sliderStyles,
   );
 
   const inputStyles = classnames(styles.checkbox, {
@@ -86,12 +103,12 @@ export default function Switch({
         disabled={disabled}
         id={id}
         name={name}
-        onBlur={() => setFocused(false)}
+        onBlur={handleOnBlur}
         onChange={handleChange}
-        onFocus={() => setFocused(true)}
+        onFocus={handleOnFocus}
         type="checkbox"
       />
-      <div className={sliderStyles} />
+      <div className={isInVRExperiment ? sliderVrStyles : sliderStyles} />
     </div>
   );
 }